chore: Move readJsonFile from text/tools.js to path/tools.js
