+++ lib/importer.js
@@ -4,7 +4,7 @@ import stripMarkdown from "strip-markdown";
 import { unified } from "unified";
 import { VFile } from "vfile";
 import { ExportTypeGlossary } from "./model/export-type-glossary.js";
-import { readJsonFile } from "./text/tools.js";
+import { readJsonFile } from "./path/tools.js";
 
 const jsonLdModuleLoaded = import("jsonld")
     .then(module => module.default)
