+++ lib/path/tools.js
@@ -1,4 +1,5 @@
-import fs from "node:fs";
+import fs, { readFile } from "node:fs";
+import https from "node:https";
 import path from "node:path";
 import url from "node:url";
 import { VFile } from "vfile";
@@ -215,4 +216,104 @@ export function getUrlPath(from, to) {
     return toForwardSlash(path.resolve(from_, to_))
         .replace(from_, "")
         .replace(LEADING_SLASHES, "$2");
+}
+
+
+/**
+ * Reads and parses a JSON file from a local path (relative to baseDir) or from
+ * an https:// or file:// URL. Other URL schemes (including http://) are not
+ * supported.
+ *
+ * @param {string} jsonFileNameOrUrl (only https:// or file://)
+ * @param {string} baseDir
+ * @returns Promise<any>
+ */
+ export function readJsonFile(jsonFileNameOrUrl, baseDir) {
+    let promise;
+    const fileScheme = "file://";
+    const httpScheme = "http://";
+    const httpsScheme = "https://";
+    if (jsonFileNameOrUrl.substr(0, httpsScheme.length) === httpsScheme) {
+        // https://
+        promise = readJsonFileUrl(jsonFileNameOrUrl);
+    } else if (jsonFileNameOrUrl.substr(0, httpScheme.length) === httpScheme) {
+        // http://
+        promise = Promise.reject(`Reading files from ${httpScheme} URL not supported. Use ${httpsScheme} or ${fileScheme} URLs or filesystem paths.`);
+    } else {
+        // file:// and none...
+        promise = new Promise((resolve, reject) => {
+            const cb = (err, jsonStr) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    try {
+                        resolve(jsonStr);
+                    } catch (parseErr) {
+                        reject(parseErr);
+                    }
+                }
+            };
+            if (jsonFileNameOrUrl.substr(0, fileScheme.length) === fileScheme) {
+                readFile(jsonFileNameOrUrl, cb);
+            } else {
+                readFile(path.resolve(baseDir, jsonFileNameOrUrl), cb);
+            }
+        });
+    }
+    return promise.then(jsonStr => JSON.parse(jsonStr));
+}
+
+/*
+ * Content Types supported to be read from a remote URL. For up to 10 content
+ * types build a quality-weighted HTTP Accept header string
+ * ~~~
+ * mime/type1, mime/type2;q=0.9, mime/type3;q=0.8, ...
+ * ~~~
+ */
+const contentTypes = ["application/json", "application/ld+json", "text/plain"];
+const contentTypesTest = new RegExp(`(${contentTypes.join("|")})`);
+const contentTypesAccept = contentTypes
+    .slice(0, 9)
+    .map((type, i) => `${type}${i > 0 ? ";q=" + (1.0 - i * 0.1) : ""}`)
+    .join(", ");
+
+function readJsonFileUrl(jsonFileUrl) {
+    return new Promise((resolve, reject) => {
+        let url;
+        try {
+            url = new URL(jsonFileUrl);
+        } catch (err) {
+            return reject();
+        }
+        try {
+            console.log(`GET ${url.toString()}`);
+            let jsonStr = "";
+            const req = https
+                .get(url, {
+                    headers: {
+                        "Accept": contentTypesAccept
+                        ,"Accept-Charset": "utf-8"
+                        ,"Cache-Control": "no-cache"
+                    }
+                }, (res) => {
+                    const contentType = res.headers["content-type"] || res.headers["Content-Type"] || "";
+                    if (! contentType.match(contentTypesTest)) {
+                        req.destroy(`Invalid response. Expecting "Content-Type" header matching expression "${contentTypesTest}" but got "${contentType}".`);
+                        reject(`Invalid response. Expecting "Content-Type" header matching expression "${contentTypesTest}" but got "${contentType}".`);
+                    }
+                    res.on("data", chunk => jsonStr += chunk);
+                    res.on("end", () => {
+                        console.log("Finished.");
+                        resolve(jsonStr);
+                    });
+                    res.on("error", reject);
+                })
+                .on("error", reject);
+            req.setTimeout(30000, () => {
+                reject("Connection timed out.");
+            });
+        } catch (err) {
+            reject(err);
+        }
+    });
 }
